+++ tests/test_moldrug.py
@@ -340,4 +340,5 @@ def test_constraintconf():
 
 
 if __name__ == '__main__':
+    test_single_receptor_command_line()
     pass